% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_generate.R
\name{alt_generate}
\alias{alt_generate}
\title{Generate source code for an altrep wrapper}
\usage{
alt_generate(altrep_type)
}
\arguments{
\item{altrep_type}{A specific type of altrep vector; 'integer', 'real' or 'logical'}
}
\value{
a wrapper around the altrep vector which is an altrep vector itself
}
\description{
Generate source code for an altrep wrapper
}
