% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_wrap.R
\name{alt_wrap}
\alias{alt_wrap}
\title{Wrap an altrep vector}
\usage{
alt_wrap(altrep_vec)
}
\arguments{
\item{altrep_vec}{altrep vector}
}
\value{
a wrapper around the altrep vector which is an altrep vector itself
}
\description{
Wrap an altrep vector
}
