% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_wrap_integer.R
\name{alt_wrap_integer}
\alias{alt_wrap_integer}
\title{Wrap an integer altrep vector}
\usage{
alt_wrap_integer(altrep_vec, vec_id)
}
\arguments{
\item{altrep_vec}{altrep vector}
}
\value{
a wrapper around the altrep vector which is an altrep vector itself
}
\description{
Wrap an integer altrep vector
}
