% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altvec_helpers.R
\name{altrep_class}
\alias{altrep_class}
\title{Retrieve the ALTREP class descriptor}
\usage{
altrep_class(x)
}
\arguments{
\item{x}{ALTREP vector for which to determine the ALTREP class descriptor}
}
\value{
ALTREP class descriptor
}
\description{
Retrieve the ALTREP class descriptor
}
