% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altrep_helpers.R
\name{altrep_data}
\alias{altrep_data}
\title{Retireve the internal ALTREP representation}
\usage{
altrep_data(x)
}
\arguments{
\item{x}{ALTREP vector for which to determine the ALTREP class descriptor}
}
\value{
internal ALTREP representation
}
\description{
Retireve the internal ALTREP representation
}
