% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/native_vec.R
\name{altrep_dataptr.native_vec}
\alias{altrep_dataptr.native_vec}
\title{Generic function to emulate the altrep dataptr method}
\usage{
\method{altrep_dataptr}{native_vec}(x, ...)
}
\arguments{
\item{x}{a state object}

\item{...}{}
}
\value{
result of the altrep dataptr method
}
\description{
Generic function to emulate the altrep dataptr method
}
