% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_vec_generics.R
\name{altrep_dataptr}
\alias{altrep_dataptr}
\title{Generic function to emulate the altrep dataptr method}
\usage{
altrep_dataptr(x, ...)
}
\arguments{
\item{x}{a state object}

\item{...}{}
}
\value{
result of the altrep dataptr method
}
\description{
Generic function to emulate the altrep dataptr method
}
