% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/native_vec.R
\name{altrep_element.native_vec}
\alias{altrep_element.native_vec}
\title{Generic function to emulate the altrep element method}
\usage{
\method{altrep_element}{native_vec}(x, ...)
}
\arguments{
\item{x}{a state object}

\item{...}{additional arguments, ignored}
}
\value{
result of the altrep element method
}
\description{
Generic function to emulate the altrep element method
}
