% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_vec_generics.R
\name{altrep_element}
\alias{altrep_element}
\title{Generic function to emulate the altrep element method}
\usage{
altrep_element(x, ...)
}
\arguments{
\item{x}{a state object}

\item{...}{}
}
\value{
result of the altrep element method
}
\description{
Generic function to emulate the altrep element method
}
