% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/native_vec.R
\name{altrep_get_region.native_vec}
\alias{altrep_get_region.native_vec}
\title{Generic function to emulate the altrep get_region method}
\usage{
\method{altrep_get_region}{native_vec}(x, ...)
}
\arguments{
\item{x}{a state object}

\item{...}{additional arguments, ignored}
}
\value{
result of the altrep get_region method
}
\description{
Generic function to emulate the altrep get_region method
}
