% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_vec_generics.R
\name{altrep_get_region}
\alias{altrep_get_region}
\title{Generic function to emulate the altrep get_region method}
\usage{
altrep_get_region(x, ...)
}
\arguments{
\item{x}{a state object}

\item{...}{}
}
\value{
result of the altrep get_region method
}
\description{
Generic function to emulate the altrep get_region method
}
