% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_vec_generics.R
\name{altrep_inspect}
\alias{altrep_inspect}
\title{Generic function to emulate the altrep inspect method}
\usage{
altrep_inspect(x, ...)
}
\arguments{
\item{x}{a state object}

\item{...}{}
}
\value{
result of the altrep inspect method
}
\description{
Generic function to emulate the altrep inspect method
}
