% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{altrep_inspect_method}
\alias{altrep_inspect_method}
\title{Trigger ALTREP_INSPECT method}
\usage{
altrep_inspect_method(x, pre, deep, pvec)
}
\arguments{
\item{x}{is the object to inspect}

\item{pre}{is the prefix}

\item{deep}{specifies the recursion behavior (0 = no recursion, -1 = [sort of] unlimited
recursion, positive numbers define the maximum recursion depth)}

\item{pvec}{is the maximum number of vector elements to show}
}
\description{
Trigger ALTREP_INSPECT method
}
