% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/native_vec.R
\name{altrep_length.native_vec}
\alias{altrep_length.native_vec}
\title{Generic function to emulate the altrep length method}
\usage{
\method{altrep_length}{native_vec}(x, ...)
}
\arguments{
\item{x}{a state object}

\item{...}{additional arguments, ignored}
}
\value{
integer result of the altrep length method
}
\description{
Generic function to emulate the altrep length method
}
