% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altrep_helpers.R
\name{altrep_length}
\alias{altrep_length}
\title{Trigger ALTREP method altrep_Length}
\usage{
altrep_length(x)
}
\arguments{
\item{x}{ALTREP vector}
}
\value{
value returned by ALTREP method altrep_Length
}
\description{
Trigger ALTREP method altrep_Length
}
