% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altrep_listener.R
\name{altrep_listener}
\alias{altrep_listener}
\title{Add listeners to an ALTREP vector to eavesdrop on all internal ALTREP calls}
\usage{
altrep_listener(altrep_vec, vec_id)
}
\arguments{
\item{altrep_vec}{altrep vector}

\item{vec_id}{custom id of the ALTREP vector, used in diagnostic output}
}
\description{
This method adds a set of listeners to an ALTREP vector that can be used to
track all internal ALTREP calls. Diagnostic information is shown on the specific
calls that are made from the R language and the data returned from the ALTREP vector.
}
