% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altrep_helpers.R
\name{altrep_package}
\alias{altrep_package}
\title{Retrieve the package in which a ALTREP vector is defined}
\usage{
altrep_package(x)
}
\arguments{
\item{x}{ALTREP vector for which to determine the package}
}
\value{
package that contains the ALTREP class definition
}
\description{
Retrieve the package in which a ALTREP vector is defined
}
