% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/native_vec.R
\name{altrep_serialized_state.native_vec}
\alias{altrep_serialized_state.native_vec}
\title{Generic function to emulate the altrep unserialize method}
\usage{
\method{altrep_serialized_state}{native_vec}(x, ...)
}
\arguments{
\item{x}{a state object}

\item{...}{}
}
\value{
result of the altrep unserialize method
}
\description{
Generic function to emulate the altrep unserialize method
}
