% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_vec_generics.R
\name{altrep_serialized_state}
\alias{altrep_serialized_state}
\title{Generic function to emulate the altrep unserialize method}
\usage{
altrep_serialized_state(x, ...)
}
\arguments{
\item{x}{a state object}

\item{...}{}
}
\value{
result of the altrep unserialize method
}
\description{
Generic function to emulate the altrep unserialize method
}
