% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altrep_helpers.R
\name{altrep_type}
\alias{altrep_type}
\title{ALTREP base type}
\usage{
altrep_type(x)
}
\arguments{
\item{x}{ALTREP vector for which to determine the type}
}
\value{
ALTREP vector base type
}
\description{
ALTREP base type
}
