% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altrep_wrapper.R
\name{altrep_wrapper}
\alias{altrep_wrapper}
\title{Wrap an altrep vector}
\usage{
altrep_wrapper(altrep_vec)
}
\arguments{
\item{altrep_vec}{altrep vector}
}
\value{
a wrapper around the altrep vector which is an altrep vector itself
}
\description{
Wrap an altrep vector
}
