% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyvec_diagnostics.R
\name{diagnostic_coerce}
\alias{diagnostic_coerce}
\title{Coerce vector to a different type}
\usage{
diagnostic_coerce(x, type)
}
\arguments{
\item{x}{vector}

\item{type}{type to coerce to. Possible types are:}
}
\value{
coerced vector (which can  be an ALTREP)
}
\description{
Coerce vector to a different type
}
