% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyvec_helpers.R
\name{diagnostics}
\alias{diagnostics}
\title{Diagnostic methods to check boundaries and validity of return values of the
user lazyvecdefined API,}
\usage{
diagnostics()
}
\value{
list of diagnostic functions
}
\description{
Diagnostic methods to check boundaries and validity of return values of the
user lazyvecdefined API,
}
