% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altvec_helpers.R
\name{is_altrep}
\alias{is_altrep}
\title{Test if object has altrep representation}
\usage{
is_altrep(x)
}
\arguments{
\item{x}{the object to be tested}
}
\value{
TRUE if the object is an altrep vector
}
\description{
Test if object has altrep representation
}
