% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyvec.R
\name{lazyvec}
\alias{lazyvec}
\title{Define a custom ALTREP vector}
\usage{
lazyvec(altrep_vec, vec_id)
}
\arguments{
\item{altrep_vec}{altrep vector}

\item{vec_id}{id of the custom ALTREP vector}
}
\value{
a user-defined ALTREP vector
}
\description{
This method creates a user-defined ALTREP vector with the complete ALTREP API created
in R instead of C/C++ sources.
}
