% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyvec_methods.R
\name{lazyvec_methods}
\alias{lazyvec_methods}
\title{Define the lazyvec interface}
\usage{
lazyvec_methods(lazyvec_init, method_length, method_dataptr_or_null,
  method_get_region, method_element, method_dataptr, method_is_sorted,
  method_no_na, method_sum, method_min, method_max, method_inspect,
  method_unserialize_ex, method_serialized_state, method_duplicate_ex,
  method_coerce, method_extract_subset)
}
\arguments{
\item{lazyvec_init}{method to be called once upon deserialization (copying?) of a vector}

\item{method_length}{override for the ALTREP length method}

\item{method_dataptr_or_null}{override for the ALTREP dataptr_or_null method}

\item{method_get_region}{override for the ALTREP get_region method}

\item{method_element}{override for the ALTREP element method}

\item{method_dataptr}{override for the ALTREP dataptr method}

\item{method_is_sorted}{override for the ALTREP is_sorted method}

\item{method_no_na}{override for the ALTREP no_na method}

\item{method_sum}{override for the ALTREP sum method}

\item{method_min}{override for the ALTREP min method}

\item{method_max}{override for the ALTREP max method}

\item{method_inspect}{override for the ALTREP inspect method}

\item{method_unserialize_ex}{override for the ALTREP unserialize_ex method}

\item{method_serialized_state}{override for the ALTREP serialized_state method}

\item{method_duplicate_ex}{override for the ALTREP duplicate_ex method}

\item{method_coerce}{override for the ALTREP coerce method}

\item{method_extract_subset}{override for the ALTREP extract_subset method}
}
\value{
object of type 'lazyvec_api' defining the set of methods to use for ALTREP calls
}
\description{
With this method a complete ALTREP interface is defined using R functions. Calls to the custom
ALTREP vector will be diverted to these user functions.
}
