% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyvec_skeleton.R
\name{lazyvec_skeleton}
\alias{lazyvec_skeleton}
\title{Generate a fully functional lazyvec implementation from a selection of examples}
\usage{
lazyvec_skeleton(skeleton_id, path)
}
\arguments{
\item{skeleton_id}{id of the sample implementation. Valid id's are:}

\item{path}{path to the target file}
}
\description{
With this method you can generate a single-file sample implementation of a lazyvec.
This sample implementation can be used as a starting point for custom implementations.
For a detailed description of the sample implementations, please see the corresponding
vignette.
}
\details{
The following sample implementations are available:
\itemize{
\item \emph{integer_range}: an integer range with a custom step size
\item \emph{real_range}: a range of double's with a custom step size
\item \emph{logical_compressed}: a compressed logical vector
\item \emph{raw_ondisk}: a file based out-of-memory raw vector
\item \emph{character_range}: the character vector equivalent of an integers range n:m
}
}
