% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{lazyvec_version}
\alias{lazyvec_version}
\title{Return lazyvec version identifier}
\usage{
lazyvec_version()
}
\description{
The version is calculated as: `256 * major_version + 16 * minor_version + patch_version`
}
