% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/native_vec.R
\name{native_vec}
\alias{native_vec}
\title{Create a altrep interface object of class 'native_vec'}
\usage{
native_vec(vec)
}
\arguments{
\item{vec}{an ordinary R vector}
}
\value{
object of class 'native_vec' for which the altrep generics are implemented
}
\description{
Create a altrep interface object of class 'native_vec'
}
