% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/native_vec.R
\name{vec.altrep_inspect.native_vec}
\alias{vec.altrep_inspect.native_vec}
\title{Generic function to emulate the altrep inspect method}
\usage{
vec.altrep_inspect.native_vec(x, ...)
}
\arguments{
\item{x}{a state object}

\item{...}{}
}
\value{
result of the altrep inspect method
}
\description{
Generic function to emulate the altrep inspect method
}
