% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_wrap.R
\name{alt_wrap}
\alias{alt_wrap}
\title{Wrap an altrep vector with listeners}
\usage{
alt_wrap(altrep_vec, vec_id)
}
\arguments{
\item{altrep_vec}{altrep vector}

\item{vec_id}{id of the encapsulated vector, used in diagnostic output}
}
\value{
a wrapper around the altrep vector which in turn is an altrep vector itself
}
\description{
This method encapsulates an ALTREP vector with a new ALTREP vector that\
passes all ALTREP calls to the original vector. In addition to that, diagnostic
information is sent to the console about the specific calls executed from the R
framework and with reports on the data received from the encapsulated vector.
}
