% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_vec_generics.R
\name{altrep_element}
\alias{altrep_element}
\title{Retrieve a single element from a vector}
\usage{
altrep_element(x, i, ...)
}
\arguments{
\item{x}{a state object}

\item{i}{index of the element to retrieve}

\item{...}{additional arguments, ignored}
}
\value{
result of the altrep element method
}
\description{
Retrieve a single element from a vector
}
