% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_vec_generics.R
\name{altrep_length}
\alias{altrep_length}
\title{Generic function to emulate the altrep length method}
\usage{
altrep_length(x, ...)
}
\arguments{
\item{x}{a state object}

\item{...}{additional arguments, ignored}
}
\value{
result of the altrep length method
}
\description{
Generic function to emulate the altrep length method
}
