% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_vec.R
\name{lazy_vec}
\alias{lazy_vec}
\title{Wrapper around an altrep interface}
\usage{
lazy_vec(altrep_interface)
}
\arguments{
\item{altrep_interface}{altrep_interface object}
}
\value{
an altrep vector
}
\description{
Wrapper around an altrep interface
}
