% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_vec.R
\name{lazy_vec}
\alias{lazy_vec}
\title{Create a lazy vector}
\usage{
lazy_vec(lazyvec_interface)
}
\arguments{
\item{lazyvec_interface}{Interface class for the lazy vector}
}
\value{
Wrapper around a lazy vector
}
\description{
Create a lazy vector
}
