% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyvec.R
\name{lazyvec}
\alias{lazyvec}
\title{Create a lazy vector}
\usage{
lazyvec(lazyvec_interface)
}
\arguments{
\item{lazyvec_interface}{Interface class for the lazy vector}
}
\value{
Wrapper around a lazy vector
}
\description{
Create a lazy vector
}
