% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyvec.R
\name{lazyvec}
\alias{lazyvec}
\title{Define a custom ALTREP vector}
\usage{
lazyvec(metadata, vec_type, altrep_methods)
}
\arguments{
\item{metadata}{custom metadata stored alongside the generated ALTREP vector.}

\item{vec_type}{data type required for the ALTREP vector, options are 'integer', 'double',
'logical', 'raw' and 'character'.}

\item{altrep_methods}{user-defined methods used by the resulting ALTREP vector.}
}
\value{
a user-defined ALTREP vector
}
\description{
This method creates a user-defined ALTREP vector with the complete ALTREP API created
in R instead of C/C++ sources.
}
